+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Please mark all changes in change log and use the issue from GitHub
 -   \#2585 Support IVF_PQ on GPU with using metric_type IP
 -   \#2689 Construct Knowhere Index Without Data
 -   \#2798 hnsw-sq8 support
+-   \#2802 Add new index: IVFSQ8NR
 
 ## Improvement
 -   \#2543 Remove secondary_path related code